requirements.txt
airflow-dbt==0.3.0
awswrangler==2.8.0
boto3==1.17.72
botocore==1.20.93
elasticsearch==7.8.1
fsspec==2021.6.1
JayDeBeApi==1.2.3
mysql-connector-python==8.0.23
oauth2==1.9.0.post1
oauth2client==4.1.3
pandas==1.2.4
pyarrow==3.0.0
pydrive==1.3.1
pymssql==2.1.5
python-oauth2==1.1.1
requests==2.23.0
s3fs==0.4.0
snowflake-connector-python==2.4.1
snowflake-sqlalchemy==1.2.4
spacy==2.2.0
tableauserverclient==0.16.0
XlsxWriter==3.0.1
yake==0.4.8
apscheduler
pymongo
Django